+++ HalflingEnemy.cc
@@ -4,7 +4,7 @@
 HalflingEnemy::HalflingEnemy(size_t x, size_t y) : Enemy(x,y,'L', 100, 100, 15, 20){}
 
 //50% chance to dodge
-void HalflingEnemy::damage(Character& c, int damage){
+void HalflingEnemy::damage(Character& c, double damage){
 	int roll = (std::rand() % 2);
 	if (roll == 0){
 		//dodge
