+++ Merchant.cc
@@ -2,7 +2,7 @@
 
 Merchant::Merchant(size_t x, size_t y) : Enemy(x,y,'M', 30, 30, 70, 5){}
 
-void Merchant::damage(Character &c, int damage){
+void Merchant::damage(Character &c, double damage){
 	//TODO Make all merchants hostile
 	BaseCharacter::damage(c, damage);
 }
